use int in progress bar

+++ chess_bingo_app/app.py
@@ -84,7 +84,7 @@ def get_user_data(username,defaultusername='khg002'):
         response = requests.get(url,headers=headers,stream=True)
         print('received response from lichess api')
         for chunk in response.iter_content(chunk_size=1024):
-            percentage_complete = (chunks / chunks_expected) * 100
+            percentage_complete = int((chunks / chunks_expected) * 100)
             socketio.emit('progress', percentage_complete)
             chunks += 1
             print(percentage_complete)
